Update of Bluewing Server to build 17, Client to build 84
Update of Bluewing Server to build 17. Includes fixes from commits of July 5th, 2020 - July 6th, 2020, from SortaCore/bluewing-cpp-server , for example SortaCore/bluewing-cpp-server@0ac3f733b6f476a544fc12f29b3b87c08fef3872.
Build 17 is the final internally ANSI build of Bluewing Server; all following builds will be TCHAR/UTF-8 based, with both ANSI and Unicode builds.
Notable fixes are connections no longer being accepted after an indefinite time being fixed, and memory leaks fixed, with a clean shutdown even in the middle of high activity. Also some prep for other platform ports. For the long list of further fixes, refer to 0ac3f733b6f476a544fc12f29b3b87c08fef3872.

Bluewing Client updated to build 84. Its changes were more a side-effect of the things it used; i.e. by virtue of liblacewing updating from Server fixes, DarkEdif SDK being minimised, etc.

+++ DarkEdif/Lib/Shared/Lacewing/FrameReader.h
@@ -52,7 +52,7 @@ public:
 	framereader() {
 	}
 
-	inline void process(const char * data, unsigned int size)
+	inline void process(const char * data, size_t size)
 	{
 		while(state < 3 && size -- > 0)
 		{
@@ -163,7 +163,7 @@ public:
 			}
 		}
 
-		unsigned int thismessagebytes = messagesize - buffer.size;
+		size_t thismessagebytes = messagesize - buffer.size;
 
 		if (size < thismessagebytes)
 			thismessagebytes = size;
